# alon_sentinel/docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Alon Sentinel API
  version: 1.0.0
  summary: Stable public HTTP contract for Alon Sentinel v1.
  description: |
    Alon Sentinel is an API-first monitoring service. This contract mirrors
    the Axum route surface implemented under `src/api/routes`.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/token:
    post:
      operationId: issueApiToken
      tags: [Auth]
      summary: Issue an API client access token
      security: []
      requestBody: { $ref: "#/components/requestBodies/IssueToken" }
      responses: { "200": { $ref: "#/components/responses/JsonOk" }, "400": { $ref: "#/components/responses/BadRequest" }, "401": { $ref: "#/components/responses/Unauthorized" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
...
  requestBodies:
    IssueToken: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/IssueTokenRequest" } } } }
...
  responses:
    JsonOk:
      description: Success
      content: { application/json: { schema: { $ref: "#/components/schemas/JsonValue" } } }
...
    BadRequest: { description: Bad request, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Unauthorized: { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  schemas:
    JsonValue: {}
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { type: string } }
    IssueTokenRequest:
      type: object
      required: [client_id, client_secret]
      properties: { client_id: { type: string }, client_secret: { type: string, writeOnly: true } }
...